+++ CAT/workflows/workflow_dicts.py
@@ -7,73 +7,83 @@ A module for loading :class:`WorkFlow` templates.
 """
 
 import os
+from types import MappingProxyType
+from typing import Mapping, MutableMapping, Tuple
 
 import yaml
 import numpy as np
 
-from CAT.frozen_settings import FrozenSettings
+from .key_map import (
+    OPT,
+    MOL,
+    HDF5_INDEX,
+    JOB_SETTINGS_QD_OPT,
+    JOB_SETTINGS_CRS,
+    JOB_SETTINGS_BDE,
+    JOB_SETTINGS_ASA,
+    ASA_INT,
+    ASA_STRAIN,
+    ASA_E,
+    SETTINGS1,
+    SETTINGS2,
+    SETTINGS_SOLV1,
+    SETTINGS_SOLV2,
+    SETTINGS_ASA,
+    SETTINGS_BDE1,
+    SETTINGS_BDE2,
+    V_BULK
+)
 
-__all__ = ['finalize_templates']
-
-# TODO import this stuff from a module
-OPT = ('opt', '')
-MOL = ('mol', '')
-HDF5_INDEX = ('hdf5 index', '')
-JOB_SETTINGS_QD_OPT = ('job_settings_qd_opt', '')
-JOB_SETTINGS_CRS = ('job_settings_crs', '')
-JOB_SETTINGS_BDE = ('job_settings_BDE', '')
-JOB_SETTINGS_ASA = ('job_settings_ASA', '')
-ASA_INT = ('ASA', 'E_int')
-ASA_STRAIN = ('ASA', 'E_strain')
-ASA_E = ('ASA', 'E')
-SETTINGS1 = ('settings', '1')
-SETTINGS2 = ('settings', '2')
-SETTINGS_SOLV1 = ('settings', 'solv 1')
-SETTINGS_SOLV2 = ('settings', 'solv 2')
-SETTINGS_ASA = ('settings', 'ASA 1')
-SETTINGS_BDE1 = ('settings', 'BDE 1')
-SETTINGS_BDE2 = ('settings', 'BDE 2')
-V_BULK = ('V_bulk', '')
-
-
-#: A dictionary with additional values for ``CAT/data/workflow_dicts/workflow_yaml.yaml``.
-#: Contains values which are generally not as easily represented in the .yaml format.
-_TEMPLATE_UPDATE: dict = {
-    'asa': {'import_columns': {ASA_INT: np.nan, ASA_STRAIN: np.nan, ASA_E: np.nan},
-            'export_columns': (JOB_SETTINGS_ASA, SETTINGS_ASA, ASA_INT, ASA_STRAIN, ASA_E)},
-    'ligand_opt': {'import_columns': {HDF5_INDEX: -1, OPT: False},
-                   'export_columns': (HDF5_INDEX, OPT, SETTINGS1, SETTINGS2)},
-    'qd_attach': {'import_columns': {HDF5_INDEX: -1, OPT: False},
-                  'export_columns': (HDF5_INDEX,)},
-    'qd_opt': {'import_columns': {HDF5_INDEX: -1, OPT: False},
-               'export_columns': (HDF5_INDEX, OPT, JOB_SETTINGS_QD_OPT, SETTINGS1, SETTINGS2)},
-    'crs': {'import_columns': {},
-            'export_columns': (JOB_SETTINGS_CRS, SETTINGS_SOLV1, SETTINGS_SOLV2)},
-    'bde': {'import_columns': {},
-            'export_columns': (JOB_SETTINGS_BDE, SETTINGS_BDE1, SETTINGS_BDE2)},
-    'forcefield': {'import_columns': {},
-                   'export_columns': ()},
-    'bulkiness': {'import_columns': {V_BULK: np.nan},
-                  'export_columns': (V_BULK,)}
-}
+__all__ = ['WORKFLOW_TEMPLATE']
 
 
-def load_templates() -> dict:
+def _load_templates() -> dict:
     """Load the templates from ``CAT/data/workflow_dicts/workflow_yaml.yaml``."""
     path = os.path.join(os.path.dirname(__file__), 'workflow_yaml.yaml')
     with open(path, 'r') as f:
         return yaml.load(f, Loader=yaml.FullLoader)
 
 
-def finalize_templates() -> FrozenSettings:
+def _recursive_mapping_proxy(dct: MutableMapping) -> MappingProxyType:
+    """Recursivelly convert **dct**, and all nested mutable mappings, into :class:`MappingProxyType<types.MappingProxyType>` instances."""  # noqa
+    for k, v in dct.items():
+        if isinstance(v, MutableMapping):
+            dct[k] = _recursive_mapping_proxy(v)
+    return MappingProxyType(dct)
+
+
+def finalize_templates():
     """Update the templates using :data:`._TEMPLATE_UPDATE`."""
-    templates = load_templates()
+    base_dict = {
+        'asa': {'import_columns': {ASA_INT: np.nan, ASA_STRAIN: np.nan, ASA_E: np.nan},
+                'export_columns': (JOB_SETTINGS_ASA, SETTINGS_ASA, ASA_INT, ASA_STRAIN, ASA_E)},
+        'ligand_opt': {'import_columns': {HDF5_INDEX: -1, OPT: False},
+                       'export_columns': (HDF5_INDEX, OPT, SETTINGS1, SETTINGS2)},
+        'qd_attach': {'import_columns': {HDF5_INDEX: -1, OPT: False},
+                      'export_columns': (HDF5_INDEX,)},
+        'qd_opt': {'import_columns': {HDF5_INDEX: -1, OPT: False},
+                   'export_columns': (HDF5_INDEX, OPT, JOB_SETTINGS_QD_OPT, SETTINGS1, SETTINGS2)},
+        'crs': {'import_columns': {},
+                'export_columns': (JOB_SETTINGS_CRS, SETTINGS_SOLV1, SETTINGS_SOLV2)},
+        'bde': {'import_columns': {},
+                'export_columns': (JOB_SETTINGS_BDE, SETTINGS_BDE1, SETTINGS_BDE2)},
+        'forcefield': {'import_columns': {},
+                       'export_columns': ()},
+        'bulkiness': {'import_columns': {V_BULK: np.nan},
+                      'export_columns': (V_BULK,)}
+    }
+
+    templates = _load_templates()
     for k, v1 in templates.items():
-        v2 = _TEMPLATE_UPDATE[k]
+        v2 = base_dict[k]
         v1.update(v2)
 
         # Convert all lists stored under the 'template' key into tuples
         for key, list_ in v1['template'].items():
             v1['template'][key] = tuple(list_)
+    return _recursive_mapping_proxy(templates)
 
-    return FrozenSettings(templates)
+
+#: An immutable mapping with additional values for ``CAT/data/workflow_dicts/workflow_yaml.yaml``.
+#: Contains values which are generally not as easily represented in the .yaml format.
+WORKFLOW_TEMPLATE: Mapping[str, Mapping] = finalize_templates()